# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Diet tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type MealType = typeof MealType[keyof typeof MealType] | null;


export const MealType = {
  healthy: 'healthy',
  moderate: 'moderate',
  junk: 'junk',
} as const;

export type UserProfileGender = typeof UserProfileGender[keyof typeof UserProfileGender];


export const UserProfileGender = {
  male: 'male',
  female: 'female',
} as const;

export type UserProfileActivityLevel = typeof UserProfileActivityLevel[keyof typeof UserProfileActivityLevel];


export const UserProfileActivityLevel = {
  sedentary: 'sedentary',
  light: 'light',
  moderate: 'moderate',
  active: 'active',
} as const;

export interface UserProfile {
  id: number;
  weightKg: number;
  heightCm: number;
  age: number;
  gender: UserProfileGender;
  activityLevel: UserProfileActivityLevel;
  dailyCalorieGoal: number;
  dailyProteinGoal: number;
  createdAt: string;
  updatedAt: string;
}

export type UserProfileInputGender = typeof UserProfileInputGender[keyof typeof UserProfileInputGender];


export const UserProfileInputGender = {
  male: 'male',
  female: 'female',
} as const;

export type UserProfileInputActivityLevel = typeof UserProfileInputActivityLevel[keyof typeof UserProfileInputActivityLevel];


export const UserProfileInputActivityLevel = {
  sedentary: 'sedentary',
  light: 'light',
  moderate: 'moderate',
  active: 'active',
} as const;

export interface UserProfileInput {
  weightKg: number;
  heightCm: number;
  age: number;
  gender: UserProfileInputGender;
  activityLevel: UserProfileInputActivityLevel;
}

export interface CalorieEstimateInput {
  foodDescription: string;
}

export type CalorieEstimateResultMealType = typeof CalorieEstimateResultMealType[keyof typeof CalorieEstimateResultMealType];


export const CalorieEstimateResultMealType = {
  healthy: 'healthy',
  moderate: 'moderate',
  junk: 'junk',
} as const;

export type CalorieEstimateResultItemsItemCategory = typeof CalorieEstimateResultItemsItemCategory[keyof typeof CalorieEstimateResultItemsItemCategory];


export const CalorieEstimateResultItemsItemCategory = {
  healthy: 'healthy',
  moderate: 'moderate',
  junk: 'junk',
} as const;

export type CalorieEstimateResultItemsItem = {
  food: string;
  calories: number;
  proteinG: number;
  category: CalorieEstimateResultItemsItemCategory;
};

export interface CalorieEstimateResult {
  estimatedCalories: number;
  estimatedProteinG: number;
  mealType: CalorieEstimateResultMealType;
  items: CalorieEstimateResultItemsItem[];
}

export type DietLogDayStatus = typeof DietLogDayStatus[keyof typeof DietLogDayStatus];


export const DietLogDayStatus = {
  clean: 'clean',
  cheat: 'cheat',
  moderate: 'moderate',
  empty: 'empty',
} as const;

export interface DietLog {
  id: number;
  date: string;
  breakfastType: MealType | null;
  lunchType: MealType | null;
  dinnerType: MealType | null;
  snacksType: MealType | null;
  /** @nullable */
  breakfastFood: string | null;
  /** @nullable */
  lunchFood: string | null;
  /** @nullable */
  dinnerFood: string | null;
  /** @nullable */
  snacksFood: string | null;
  waterCups: number;
  /** @nullable */
  calories: number | null;
  /** @nullable */
  protein: number | null;
  /** @nullable */
  note: string | null;
  dayStatus: DietLogDayStatus;
  createdAt: string;
  updatedAt: string;
}

export interface DietLogInput {
  date: string;
  breakfastType?: MealType | null;
  lunchType?: MealType | null;
  dinnerType?: MealType | null;
  snacksType?: MealType | null;
  /** @nullable */
  breakfastFood?: string | null;
  /** @nullable */
  lunchFood?: string | null;
  /** @nullable */
  dinnerFood?: string | null;
  /** @nullable */
  snacksFood?: string | null;
  waterCups?: number;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface MonthSummary {
  year: number;
  month: number;
  totalLogged: number;
  cleanDays: number;
  cheatDays: number;
  moderateDays: number;
  currentStreak: number;
  avgWaterCups: number;
  /** @nullable */
  avgCalories: number | null;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListDietLogsParams = {
year: number;
month: number;
};

export type GetMonthSummaryParams = {
year: number;
month: number;
};

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};