# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Diet tracking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: diet-logs
    description: Daily diet log operations
  - name: profile
    description: User profile operations
  - name: calories
    description: Calorie estimation
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: No profile yet
    post:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /calories/estimate:
    post:
      operationId: estimateCalories
      tags: [calories]
      summary: Estimate calories from a food description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalorieEstimateInput"
      responses:
        "200":
          description: Estimated calories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalorieEstimateResult"

  /diet-logs:
    get:
      operationId: listDietLogs
      tags: [diet-logs]
      summary: List all diet logs for a month
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of diet logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DietLog"
    post:
      operationId: upsertDietLog
      tags: [diet-logs]
      summary: Create or update a diet log for a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DietLogInput"
      responses:
        "200":
          description: Diet log saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietLog"

  /diet-logs/summary:
    get:
      operationId: getMonthSummary
      tags: [diet-logs]
      summary: Get monthly summary stats
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthSummary"

  /diet-logs/{date}:
    get:
      operationId: getDietLog
      tags: [diet-logs]
      summary: Get diet log for a specific date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diet log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietLog"
        "404":
          description: Not found

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    MealType:
      type: ["string", "null"]
      enum: [healthy, moderate, junk, null]

    UserProfile:
      type: object
      required: [id, weightKg, heightCm, age, gender, activityLevel, dailyCalorieGoal, dailyProteinGoal, createdAt, updatedAt]
      properties:
        id:
          type: integer
        weightKg:
          type: number
        heightCm:
          type: number
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active]
        dailyCalorieGoal:
          type: integer
        dailyProteinGoal:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    UserProfileInput:
      type: object
      required: [weightKg, heightCm, age, gender, activityLevel]
      properties:
        weightKg:
          type: number
        heightCm:
          type: number
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active]

    CalorieEstimateInput:
      type: object
      required: [foodDescription]
      properties:
        foodDescription:
          type: string

    CalorieEstimateResult:
      type: object
      required: [estimatedCalories, estimatedProteinG, mealType, items]
      properties:
        estimatedCalories:
          type: integer
        estimatedProteinG:
          type: integer
        mealType:
          type: string
          enum: [healthy, moderate, junk]
        items:
          type: array
          items:
            type: object
            required: [food, calories, proteinG, category]
            properties:
              food:
                type: string
              calories:
                type: integer
              proteinG:
                type: integer
              category:
                type: string
                enum: [healthy, moderate, junk]

    DietLog:
      type: object
      required: [id, date, breakfastType, lunchType, dinnerType, snacksType, breakfastFood, lunchFood, dinnerFood, snacksFood, waterCups, calories, protein, note, dayStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        date:
          type: string
        breakfastType:
          $ref: "#/components/schemas/MealType"
        lunchType:
          $ref: "#/components/schemas/MealType"
        dinnerType:
          $ref: "#/components/schemas/MealType"
        snacksType:
          $ref: "#/components/schemas/MealType"
        breakfastFood:
          type: ["string", "null"]
        lunchFood:
          type: ["string", "null"]
        dinnerFood:
          type: ["string", "null"]
        snacksFood:
          type: ["string", "null"]
        waterCups:
          type: integer
        calories:
          type: ["integer", "null"]
        protein:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
        dayStatus:
          type: string
          enum: [clean, cheat, moderate, empty]
        createdAt:
          type: string
        updatedAt:
          type: string

    DietLogInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
        breakfastType:
          $ref: "#/components/schemas/MealType"
        lunchType:
          $ref: "#/components/schemas/MealType"
        dinnerType:
          $ref: "#/components/schemas/MealType"
        snacksType:
          $ref: "#/components/schemas/MealType"
        breakfastFood:
          type: ["string", "null"]
        lunchFood:
          type: ["string", "null"]
        dinnerFood:
          type: ["string", "null"]
        snacksFood:
          type: ["string", "null"]
        waterCups:
          type: integer
        calories:
          type: ["integer", "null"]
        protein:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]

    MonthSummary:
      type: object
      required: [year, month, totalLogged, cleanDays, cheatDays, moderateDays, currentStreak, avgWaterCups, avgCalories]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalLogged:
          type: integer
        cleanDays:
          type: integer
        cheatDays:
          type: integer
        moderateDays:
          type: integer
        currentStreak:
          type: integer
        avgWaterCups:
          type: number
        avgCalories:
          type: ["number", "null"]

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string