# lib/api-zod/src/generated/types/calorieEstimateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Diet tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { CalorieEstimateResultItemsItem } from './calorieEstimateResultItemsItem';
import type { CalorieEstimateResultMealType } from './calorieEstimateResultMealType';

export interface CalorieEstimateResult {
  estimatedCalories: number;
  estimatedProteinG: number;
  mealType: CalorieEstimateResultMealType;
  items: CalorieEstimateResultItemsItem[];
}