# lib/api-zod/src/generated/types/dietLogInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Diet tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { MealType } from './mealType';

export interface DietLogInput {
  date: string;
  breakfastType?: MealType | null;
  lunchType?: MealType | null;
  dinnerType?: MealType | null;
  snacksType?: MealType | null;
  /** @nullable */
  breakfastFood?: string | null;
  /** @nullable */
  lunchFood?: string | null;
  /** @nullable */
  dinnerFood?: string | null;
  /** @nullable */
  snacksFood?: string | null;
  waterCups?: number;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  note?: string | null;
}