# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Diet tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileActivityLevel } from './userProfileActivityLevel';
import type { UserProfileGender } from './userProfileGender';

export interface UserProfile {
  id: number;
  weightKg: number;
  heightCm: number;
  age: number;
  gender: UserProfileGender;
  activityLevel: UserProfileActivityLevel;
  dailyCalorieGoal: number;
  dailyProteinGoal: number;
  createdAt: string;
  updatedAt: string;
}